refactor(manifest): Break up "Manifest" and "Manifest Provider"
